Adding gps support

+++ app/src/main/java/com/example/myapplication/DataPoint.java
@@ -2,8 +2,8 @@ package com.example.myapplication;
 
 public class DataPoint {
     long time = 0;
-    long latitude = 0;
-    long longitude = 0;
+    double latitude = 0;
+    double longitude = 0;
     float accelerometer_x = 0.0f;
     float accelerometer_y = 0.0f;
     float accelerometer_z = 0.0f;
